FinalVersion
It became clear to me that there was a bigger story that needed to be told about the testing,
and that as I worked and worked to get the data into the visual – that there is the human
toll that this testing takes (besides the obvious data). And there was no ‘pretty’ way to
show this – or to make the data cool looking or minimalist.
The human toll is that lives are affected by the radioactive fallout from the amazing
amounts of testing that all countries, but especially the U.S. and Russia had done, all in the
names of an arms race to show simply that we have the most firepower that could blow the
world up. In terms of real lives, all 8 countries tested in lower populated areas that they
have jurisdiction over, mostly in areas where indigenous people of their own countries, or
that they have relegated to isolated, desolate areas.
In 2018 we have a higher expectancy rate than we ever have, however, the expectancy rate
in these areas fall on the low end. There is also an economic disadvantage that these
indigenous populated areas have compared to the rest of the world, but this is another data
point. The simple fact that a lower life expectancy rate, also brings up the access to good
healthcare, again, another data point that could be included on this visual.